Switch to actix-web 3 (from iron) and to Rust 2018
This massive overhaul of the backend code was needed: it was not
possible to compile the project with a modern Rust toolchain for
years! As the `iron` project is dead, everything is moved to the
`active-web` framework; it brings some positive changes (true
test helpers, no more `lugh` process to spawn through our bash-
based test harness!), and some verbosity (see middlewares).