Squash migrations and use RFC3339 for dates explicitly
The new code expects RFC3339 for date, as it is the only sane format
we can easily generate from SQLite while being standard.

The previous schema used `CURRENT_TIMESTAMP` as a default value for
created_at and other time-based columns, which would render datetime
with date and time separated by a space, instead of the litteral `T`
and without any timezone data (`Z` was absent).